COPY instruction (#31)
* added missing include, added process_table to asm
* added missing clear_diags in tests
* added parselibprovider to asmprocessor
* minor fixes, made templated add_dependency wrapping cycle check
* minor fixes, unified process method in macrodef_processor
* fixed error with structuder binding captured by lambda
* moved preprocess low_language_proc method to cpp
* minor fixes
* minor fixes
* tests analyzer instantiation fix, lsp_info_proc remove fix
* added range_provider
* removed set_range from machine expression
* fill_location_counter removed const
* made space::resolve static
* minor fixes
* removed symbol_guard, moved range_provider
* antlr4 bug workaround
(cherry picked from commit 1b46e55a6f8ed4fcd8ba284391b28bf83440476a)
* antlr stability test (fails with antlr 4.7.2)
* added mnemonic checking
* replaced optional<vs_ptr*> for vs_ptr*
* fixes
* minor fixes
* Fixed asm instructions parsing bugs, edited diagnostics
* filled get_operand_value in adress_operand
* switch version to 4.7.2 with NoViableAlternative double free fix
* added to end instruction operand rule IDENTIFIER
* removed const from symbo::set_value
* added helper classes for composite_assembler_operant to get rid of cloning
* added code that recognizes whether macro library has corectly formed macro inside
* added tests
* fixed bad macro location
* added base infrastructure for copy
* added copy_statement_provider
* added copy tests and fixes
* added functionality for calling COPY from macro definition
* added support for COPY instr from lookahead
* fixed error - no loading macros in copy
* added test, fixed unbalanced macros
* removing warnings
* minor fixes, changed address offset seting to signed values
* solved stack overflow in statements with many operands
* long macro stack overflow test
* files with @ in the name were not highlighted
* @ files fix

+++ parser_library/src/context/address.cpp
@@ -11,7 +11,7 @@ using namespace hlasm_plugin::parser_library::context;
 space::space(location_counter& owner, id_index name)
 	: name(name), resolved_(false), owner_(owner) {}
 
-void space::resolve(space_ptr this_space, size_t length)
+void space::resolve(space_ptr this_space, int length)
 {
 	if (this_space->resolved_)
 		throw std::invalid_argument("space already resolved");
@@ -53,7 +53,7 @@ void space::remove_listener(address* addr)
 	listeners_.erase(it);
 }
 
-address::address(base address_base, size_t offset, const space_storage& spaces)
+address::address(base address_base, int offset, const space_storage& spaces)
 	: offset(offset)
 {
 	bases.emplace_back(address_base, 1);
